Extract 'isToString()' function to Utils
